Renamed the JtaXADatasourceBuild to the factory naming for consistency and correctness
